Refactor: extract duplicate logic
